(PC-23960)[PRO] refactor: move DeskScreen to Desk page component
