+++ pro/src/pages/Desk/BookingDetails/index.ts
