--- pro/src/screens/Desk/index.ts
@@ -1,7 +1,0 @@
-export { default as DeskScreen } from './Desk'
-export { MESSAGE_VARIANT } from './types'
-export type {
-  DeskGetBookingResponse,
-  DeskProps,
-  DeskSubmitResponse,
-} from './types'
